agentcore: the wire protocol in one module
The envelope union, the webhook reply, the presigned-URL pair, the
wake-alarm request and the forwarder's three reserved paths were each
spelled where they happened to be used — the adapter, the state sync,
the wake sink, the deploy driver — and the forwarder spelled them all
again by hand. channels/agentcore-protocol.ts is the one definition; the
forwarder, being JavaScript, is pinned to it by a test instead of an
import.
